Successfully build gridsynth with API
