+++ .gitignore
@@ -1,6 +1,8 @@
 CMake*
 *.cmake
 Makefile
+build
 *-build*
 .idea
 *.a
+*.vscode
